boards and lists debugging

+++ server/server-assets/routes/boards.js
@@ -46,6 +46,9 @@ router.put('/api/boards/:id', (req, res) =>{
 
 // Delete board
 router.delete('/api/boards/:id', (req, res) =>{
+    if(req.body.author != req.session.id) {
+        return next()
+    }
     Boards.findByIdAndRemove(req.params.id)
     .then(board=>{
         res.status(200).send({message: 'successfully deleted'})
